begginings of a new speaker
